RAG/requirements.txt
sentence-transformers
faiss-cpu
numpy
PyPDF2
